+++ core/src/com/tapiocagames/game/Food.java
@@ -7,15 +7,15 @@ import com.badlogic.gdx.graphics.Texture;
  * Created by thiago on 03/02/16.
  */
 public class Food {
-    int x;
-    int y;
+    float x;
+    float y;
     Texture texture;
 
     public Food() {
         set(0, 0, null);
     }
 
-    public void set(int x, int y, Texture texture) {
+    public void set(float x, float y, Texture texture) {
         this.x = x;
         this.y = y;
         this.texture = texture;
